% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexp.R
\name{iexp}
\alias{iexp}
\title{Intelligently take exponential}
\usage{
iexp(x, value = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{value}{A numeric indicating the value subtracted from exp(x).
Default is the smallest positive value of x.}

\item{...}{Values passed on to log, e.g. base}
}
\value{
A numeric vector, typically the logarithm of \code{x}
}
\description{
The inverse of \code{\link{ilog}} for \code{base=exp(1)}.
}
\examples{
x <- 0:2
all.equal(iexp(ilog(x)), x) # TRUE
}
\seealso{
\code{\link{exp}},\code{\link{ilog}}, \code{\link{min_value}}
}

