% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\title{Intelligently take logarithms}
\usage{
ilog(x, value = min_value(x), ...)
}
\arguments{
\item{x}{A numeric vector}

\item{value}{A numeric indicating the value to add to x before taking logs.
Default is the smallest positive value of x.}

\item{...}{Values passed on to log, e.g. base}
}
\value{
A numeric vector, typically the logarithm of \code{x}
}
\description{
Deals with zero values when taking logarithms by adding `value` to all
observations before taking logarithms. By default, `value` is the minimum
non-zero value.
}
\examples{
ilog(1:2)             # Returns log(1:2)
ilog(-2:2)            # Returns an error due to negative values
ilog(0:2)             # Returns log(0:2+1)
ilog(-2:2, value = 3) # Returns log(-2:2 + 3)
}
\seealso{
\code{\link{log}}, \code{\link{iexp}}, \code{\link{min_value}}
}
