% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_value.R
\name{min_value}
\alias{min_value}
\title{Find minimum non-zero value}
\usage{
min_value(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The minimum non-zero value of x
}
\description{
Find the minimum non-zero value in a numeric vector. Returns an error if
any negative values exist.
}
\examples{
min_value(0:2) # 1
}
\seealso{
\code{\link{ilog}}, \code{\link{iexp}}
}

